Can edit project names
Add icon for edit buttons and add these buttons to to navbar and main via
an utility function. Do some renaming. Editing uses to form for new
projects. Because of this it is necessary to track via a variable if
the form pops up because of a click on new project or on an edit button.