Render active project active
Add a field "active" and functions to set and get it to class project.
When changing current project the old project is deactivated and the
new one activated. This makes sure that there can always only exist one
active project.
DisplayController applies the CSS class active to change the color of
the active navbar button. For this to work, boardController got a
function that returns the index of a given project.
renderNavBar checks now for every project if it is active and if so
calls setActive. This is necessary because otherwise the active button
would be lost everytime the navbar is rendered anew.