Move checked status when moving todo's between projects and bugfix
Class ToDo gets a field named checked that tracks if it is checked.
It is a necessary parameter for creating new todo-items.
Displaycontroller checks for this field when rendering the todo-list
and applies the data-checked attribute if checked=true.

Fix editToDo so that it finds the right index to delete when the
project is changed (findToDoIndex needs a title, not the ToDo itself)